fix: Modify DataTable component to include default value for loading prop
This commit modifies the DataTable component to include a default value for the loading prop. Previously, the loading prop was not set to any default value, which could lead to errors when the component was rendered without the loading prop. With this change, the DataTable component can now be rendered without any errors even if the loading prop is not passed.